refactor(webviews): merge trpc.ts into appRouter.ts
Telemetry middleware, publicProcedureWithTelemetry, and the WithTelemetry helper now live in appRouter.ts alongside the router tree they configure. Per-view routers import these primitives from appRouter directly. trpc.ts is removed; five files become four.
